# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking Project Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://test-booking.test/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and get access token",
                "operationId": "login",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginRequest"
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bookings": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Store booking",
                "operationId": "storeBooking",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreBookingRequest"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Booking conflict - the user already has a booking in the selected period",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking conflict: The user already has a booking in the selected period."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource Not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/resources/{id}/bookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Get bookings for a specific resource",
                "operationId": "getBookingsByResource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resource ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource Not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bookings/{id}": {
            "delete": {
                "tags": [
                    "Booking"
                ],
                "summary": "Destroy booking by ID",
                "operationId": "destroyBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking successfully deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Get resources list",
                "operationId": "getResourcesPagination",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by resource type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter by resource name and description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedList"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Store resource",
                "operationId": "storeResource",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreResourceRequest"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginResponse": {
                "required": [
                    "user",
                    "token"
                ],
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1Qi..."
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "title": "User response",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    }
                },
                "type": "object"
            },
            "BookingResponse": {
                "title": "Booking response",
                "required": [
                    "resource_id",
                    "user_id",
                    "start",
                    "end"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "resource_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 7
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-12T10:00:00+04:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-12T12:00:00+04:00"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ResourceResponse": {
                "title": "Resource response",
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Room 101"
                    },
                    "type": {
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "type": "string",
                        "example": "Конференц-зал",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-11T14:00:00+00:00",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-12T10:30:00+00:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "errors": {
                        "type": "object",
                        "example": {
                            "start": [
                                "The start field is required."
                            ],
                            "end": [
                                "The end must be a valid date."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "PaginatedList": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ResourceResponse"
                        }
                    },
                    "meta": {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 100
                            },
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 10
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 10
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "LoginRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "email",
                                "password"
                            ],
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "user@example.com"
                                },
                                "password": {
                                    "type": "string",
                                    "format": "password",
                                    "example": "secret123"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "StoreBookingRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "resource_id",
                                "start_time",
                                "end_time"
                            ],
                            "properties": {
                                "resource_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "start_time": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2025-07-11 12:22:34"
                                },
                                "end_time": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2025-07-11 12:11:22"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "StoreResourceRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "type"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "maxLength": 255,
                                    "example": "Room A"
                                },
                                "type": {
                                    "type": "string",
                                    "maxLength": 16,
                                    "example": "room"
                                },
                                "description": {
                                    "type": "string",
                                    "maxLength": 1024,
                                    "example": "Spacious room with sea view",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Booking",
            "description": "Booking"
        },
        {
            "name": "Resources",
            "description": "Resources"
        }
    ]
}